+++ test/provider_extension_test.dart
@@ -1,5 +1,6 @@
 import 'package:catalyst_builder/catalyst_builder.dart';
 import 'package:explorator/explorator.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 
@@ -13,8 +14,21 @@ void main() {
       mockServiceProvider = MockServiceProviderForTest();
     });
 
+    void _activateKnownServices({
+      bool serviceProvider = true,
+      bool routeBuilder = true,
+      bool navigatorKey = true,
+    }) {
+      when(mockServiceProvider.has<dynamic>(ServiceProvider))
+          .thenReturn(serviceProvider);
+      when(mockServiceProvider.has<dynamic>(RouteBuilder))
+          .thenReturn(routeBuilder);
+      when(mockServiceProvider.has<dynamic>(GlobalKey<NavigatorState>))
+          .thenReturn(navigatorKey);
+    }
+
     test('useExplorator register the service provider itself', () {
-      when(mockServiceProvider.has<dynamic>(ServiceProvider)).thenReturn(false);
+      _activateKnownServices(serviceProvider: false);
 
       mockServiceProvider.useExplorator();
       verify(mockServiceProvider.has<dynamic>(ServiceProvider));
@@ -29,8 +43,7 @@ void main() {
     });
 
     test('useExplorator register route builder', () {
-      when(mockServiceProvider.has<dynamic>(ServiceProvider)).thenReturn(true);
-      when(mockServiceProvider.has<dynamic>(RouteBuilder)).thenReturn(false);
+      _activateKnownServices(routeBuilder: false);
 
       var routeBuilder = MaterialRouteBuilder();
       mockServiceProvider.useExplorator(routeBuilder: routeBuilder);
@@ -43,6 +56,27 @@ void main() {
       expect(
         captured.captured[0](mockServiceProvider),
         const TypeMatcher<RouteBuilder>(),
+      );
+      expect(captured.captured[1], const TypeMatcher<Service>());
+    });
+
+    test('useExplorator register a global key for the navigator state', () {
+      _activateKnownServices(navigatorKey: false);
+
+      var navigatorKey = GlobalKey<NavigatorState>();
+      mockServiceProvider.useExplorator(navigatorKey: navigatorKey);
+      verify(mockServiceProvider.has<dynamic>(GlobalKey<NavigatorState>));
+
+      var captured = verify(
+        mockServiceProvider.register<GlobalKey<NavigatorState>>(
+          captureAny,
+          captureAny,
+        ),
+      );
+
+      expect(
+        captured.captured[0](mockServiceProvider),
+        const TypeMatcher<GlobalKey<NavigatorState>>(),
       );
       expect(captured.captured[1], const TypeMatcher<Service>());
     });