+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Features:
 - Added the `RouteProvider.tag` constant to reduce magic symbols.
 - Added a ServiceProvider extension for registering the explorator services.
+- Added a `GlobalKey<NavigatorState>` to the ServiceProvider.
 
 ## 1.0.0